all done except some tests and ract-hook-form interaction with custom Input
