+++ src/components/LoginForm/LoginForm.jsx
@@ -1,30 +1,37 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Link } from 'react-router-dom';
-import { connect } from 'react-redux';
+import { useForm } from 'react-hook-form';
 import { authenticate } from '../../store/actions/auth';
+import { useDispatch } from 'react-redux';
+import { Link } from 'react-router-dom';
 import Input from '../Input/Input';
 import Button from '../Button/Button';
 
 import './LoginForm.scss';
 
-export const LoginForm = ({ authenticate }) => {
+export const LoginForm = () => {
   LoginForm.propTypes = {
     authenticate: PropTypes.func,
+  }
+
+  const dispatch = useDispatch();
+  const { register, handleSubmit } = useForm();
+
+  const onSubmit = (data) => {
+    const { email, password } = data;
+    console.log(email)
+    console.log(password)
+    dispatch(authenticate(email, password));
   }
 
   return (
     <form
       className='login-form'
-      onSubmit={(e) => {
-        e.preventDefault();
-        const emailInput = e.nativeEvent.target[0];
-        const passwordInput = e.nativeEvent.target[1];
-        authenticate(emailInput.value, passwordInput.value);
-      }}
+      onSubmit={handleSubmit(onSubmit)}
     >
       <h2 className='login-form__form-title'>Войти</h2>
-      <Input
+      <input
+        {...register('email')}
         className='login-form__input'
         label='Email'
         id='email'
@@ -32,7 +39,8 @@ export const LoginForm = ({ authenticate }) => {
         name='email'
         placeholder='mail@mail.ru'
       />
-      <Input
+      <input
+        {...register('password')}
         className='login-form__input'
         label='Пароль'
         id='password'
@@ -63,7 +71,4 @@ export const LoginForm = ({ authenticate }) => {
   )
 }
 
-export default connect(
-  (state) => ({isLoggedIn: state.auth.isLoggedIn}),
-  { authenticate }
-)(LoginForm);
+export default LoginForm;
